cancel(upload: Bool = false, background: Bool = false)

+++ iOSClient/Networking/NCNetworking.swift
@@ -772,7 +772,7 @@ class NCNetworking: NSObject, NKCommonDelegate {
     // sessionIdentifierBackgroundWWan: String = "com.nextcloud.session.upload.backgroundWWan"
     // sessionIdentifierBackgroundExtension: String = "com.nextcloud.session.upload.extension"
 
-    func cancel(inBackground: Bool) {
+    func cancel(upload: Bool = false, background: Bool = false) {
 
 #if !EXTENSION
         if let appDelegate = UIApplication.shared.delegate as? AppDelegate {
@@ -780,27 +780,42 @@ class NCNetworking: NSObject, NKCommonDelegate {
         }
 #endif
 
-        NextcloudKit.shared.sessionManager.cancelAllRequests()
-
-        downloadRequest.removeAll()
-        uploadRequest.removeAll()
+        let sessionManager = NextcloudKit.shared.sessionManager
+        sessionManager.session.getTasksWithCompletionHandler { dataTasks, uploadTasks, downloadTasks in
+            dataTasks.forEach {
+                $0.cancel()
+            }
+            downloadTasks.forEach {
+                $0.cancel()
+            }
+            uploadTasks.forEach {
+                if upload {
+                    $0.cancel()
+                }
+            }
+        }
 
         let metadatasDownload = NCManageDatabase.shared.getMetadatas(predicate: NSPredicate(format: "status < 0"))
         let metadatasUpload = NCManageDatabase.shared.getMetadatas(predicate: NSPredicate(format: "status > 0 AND session == %@", NextcloudKit.shared.nkCommonInstance.sessionIdentifierUpload))
         let metadatasUploadBackground = NCManageDatabase.shared.getMetadatas(predicate: NSPredicate(format: "status > 0 AND (session == %@ || session == %@)", NCNetworking.shared.sessionIdentifierBackground, NCNetworking.shared.sessionIdentifierBackgroundWWan))
 
         // DOWNLOAD
+        downloadRequest.removeAll()
         for metadata in metadatasDownload {
             utilityFileSystem.removeFile(atPath: utilityFileSystem.getDirectoryProviderStorageOcId(metadata.ocId))
             NCManageDatabase.shared.setMetadataSession(ocId: metadata.ocId, session: "", sessionError: "", sessionSelector: "", sessionTaskIdentifier: 0, status: NCGlobal.shared.metadataStatusNormal, errorCode: 0)
         }
+
         // UPLOAD
-        for metadata in metadatasUpload {
-            utilityFileSystem.removeFile(atPath: utilityFileSystem.getDirectoryProviderStorageOcId(metadata.ocId))
-            NCManageDatabase.shared.deleteMetadata(predicate: NSPredicate(format: "ocId == %@", metadata.ocId))
+        if upload {
+            uploadRequest.removeAll()
+            for metadata in metadatasUpload {
+                utilityFileSystem.removeFile(atPath: utilityFileSystem.getDirectoryProviderStorageOcId(metadata.ocId))
+                NCManageDatabase.shared.deleteMetadata(predicate: NSPredicate(format: "ocId == %@", metadata.ocId))
+            }
         }
 
-        guard inBackground else { return }
+        guard background else { return }
 
         // BACKGROUND
         for metadata in metadatasUploadBackground {
